Options write_enable and meas_diff
write enable: overriding setup to speed up measurements or measure with the very same settings a EIT device received from the Sciospec software if it was used with this before. This enables to measure in python even of some configuration features of the sciospec software are not implemented yet.

meas_diff : enables differential measurement mode as described in sciocom interface for EIT-128 documentation